fuentes del servicio factorial

+++ src/main/java/com/devhoss/entidades/Respuesta.java
@@ -2,11 +2,11 @@ package com.devhoss.entidades;
 
 public class Respuesta {
 private int numero;
-private int respuesta;
+private Long respuesta;
 private String version;
 
 
-public Respuesta(int numero, int respuesta, String version) {
+public Respuesta(int numero, Long respuesta, String version) {
 	super();
 	this.numero = numero;
 	this.respuesta = respuesta;
@@ -18,10 +18,10 @@ public int getNumero() {
 public void setNumero(int numero) {
 	this.numero = numero;
 }
-public int getRespuesta() {
+public Long getRespuesta() {
 	return respuesta;
 }
-public void setRespuesta(int respuesta) {
+public void setRespuesta(Long respuesta) {
 	this.respuesta = respuesta;
 }
 public String getVersion() {
